+++ samples/counter/desktop-composeui/src/main/kotlin/com/example/redwood/counter/desktop/main.kt
@@ -28,10 +28,10 @@ import app.cash.redwood.layout.composeui.ComposeUiRedwoodLayoutWidgetFactory
 import com.example.redwood.counter.composeui.ComposeUiWidgetFactory
 import com.example.redwood.counter.composeui.CounterTheme
 import com.example.redwood.counter.presenter.Counter
-import com.example.redwood.counter.widget.SchemaWidgetFactories
+import com.example.redwood.counter.widget.SchemaWidgetSystem
 
 fun main() {
-  val factories = SchemaWidgetFactories(
+  val widgetSystem = SchemaWidgetSystem(
     Schema = ComposeUiWidgetFactory,
     RedwoodLayout = ComposeUiRedwoodLayoutWidgetFactory(),
   )
@@ -43,7 +43,7 @@ fun main() {
       state = rememberWindowState(width = 300.dp, height = 300.dp),
     ) {
       CounterTheme {
-        RedwoodContent(factories, modifier = Modifier.padding(16.dp)) {
+        RedwoodContent(widgetSystem, modifier = Modifier.padding(16.dp)) {
           Counter()
         }
       }